+++ mobile/app/_layout.tsx
@@ -1,69 +1,47 @@
 import { DarkTheme, DefaultTheme, ThemeProvider } from '@react-navigation/native';
 import { PortalHost } from '@rn-primitives/portal';
+import * as Sentry from '@sentry/react-native';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { SplashScreen, Stack, useRouter } from 'expo-router';
+import { SplashScreen, Stack } from 'expo-router';
 import { StatusBar } from 'expo-status-bar';
 import 'react-native-reanimated';
 import '../global.css';
 
-import { IconSymbol } from '@/components/ui/icon-symbol';
-import { AppConfig } from '@/constants/app-config';
 import { AuthProvider } from '@/features/auth/auth-context';
 import { useAppTheme } from '@/hooks/use-app-theme';
-import * as Sentry from '@sentry/react-native';
-import { useEffect, useState } from 'react';
-import { Pressable, View } from 'react-native';
-import { useFonts } from 'expo-font';
+import { useSplashScreenReady } from '@/hooks/use-splash-screen-ready';
+import { useEffect, useRef } from 'react';
+import { Animated, View } from 'react-native';
 
-SplashScreen.preventAutoHideAsync();
+void SplashScreen.preventAutoHideAsync().catch((err) => {
+  Sentry.captureException(err);
+});
 
 const queryClient = new QueryClient();
 
 function AppLayout() {
-  const [fontsLoaded] = useFonts({
-    'Inter-Regular': require('@/assets/fonts/SpaceMono-Regular.ttf'),
+  const { isAppReady, setLayoutReady } = useSplashScreenReady();
 
-  })
-  const [isLayoutReady, setLayoutReady] = useState(false);
-  const [hasPassedMinDelay, setHasPassedMinDelay] = useState(false);
-  const [isAppReady, setAppReady] = useState(false);
+  const opacity = useRef(new Animated.Value(0)).current;
 
-  // Ensure splash shows for minimum time
   useEffect(() => {
-    const timer = setTimeout(() => setHasPassedMinDelay(true), AppConfig.LOADING.SPLASH_MIN_DISPLAY_MS);
-    return () => clearTimeout(timer);
-  }, []);
-
-  // Hide splash when all conditions are met
-  useEffect(() => {
-    const shouldHide = fontsLoaded && isLayoutReady && hasPassedMinDelay;
-
-    if (shouldHide) {
-      (async () => {
-        try {
-          await SplashScreen.hideAsync();
-        } catch (error) {
-          console.error('Splash screen error:', error);
-          Sentry.captureException(error);
-        } finally {
-          setAppReady(true);
-        }
-      })();
+    if (isAppReady) {
+      Animated.timing(opacity, {
+        toValue: 1,
+        duration: 350,
+        useNativeDriver: true,
+      }).start();
     }
-  }, [fontsLoaded, isLayoutReady, hasPassedMinDelay]);
+  }, [isAppReady]);
 
   return (
-    <View
-      className="flex-1"
-      onLayout={() => setLayoutReady(true)}
-      accessibilityLabel="app-root-view"
-    >
+    <View className="flex-1" onLayout={() => setLayoutReady(true)} accessibilityLabel="app-root-view">
       {isAppReady && (
-        <>
+        <Animated.View style={{ flex: 1, opacity }}>
           <Navigation />
           <StatusBar style="auto" />
           <PortalHost />
-        </>
+        </Animated.View>
       )}
     </View>
   );
@@ -92,28 +70,3 @@ export default function RootLayout() {
     </QueryClientProvider>
   );
 }
-
-export const useScreenOptions = () => {
-  const router = useRouter();
-  const { color } = useAppTheme();
-
-  const screenOptions = {
-    headerStyle: {},
-    headerTitleStyle: { fontSize: 20 },
-    headerTitleAlign: 'center' as 'center',
-    headerLeft: () => (
-      <Pressable
-        style={({ pressed }) => ({
-          opacity: pressed ? 0.7 : 1,
-          paddingVertical: 8,
-          paddingRight: 16,
-        })}
-        onPress={() => router.back()}
-      >
-        <IconSymbol name="chevron.left" size={28} color={color.text} />
-      </Pressable>
-    ),
-  };
-
-  return { screenOptions };
-};
