+++ mobile/hooks/use-screen-options.tsx
@@ -1,0 +1,29 @@
+import { IconSymbol } from '@/components/ui/icon-symbol';
+import { useRouter } from 'expo-router';
+import { Pressable } from 'react-native';
+import { useAppTheme } from './use-app-theme';
+
+export const useScreenOptions = () => {
+  const router = useRouter();
+  const { color } = useAppTheme();
+
+  const screenOptions = {
+    headerStyle: {},
+    headerTitleStyle: { fontSize: 20 },
+    headerTitleAlign: 'center' as 'center',
+    headerLeft: () => (
+      <Pressable
+        style={({ pressed }) => ({
+          opacity: pressed ? 0.7 : 1,
+          paddingVertical: 8,
+          paddingRight: 16,
+        })}
+        onPress={() => router.back()}
+      >
+        <IconSymbol name="chevron.left" size={36} color={color.text} />
+      </Pressable>
+    ),
+  };
+
+  return { screenOptions };
+};
